server/requirements.txt
Flask==2.3.3
python-dotenv==0.19.1
flask-cors==4.0.0
PyJWT==2.8.0